refactor: restructure for e2e tests
