+++ src/lib.rs
@@ -1,3 +1,5 @@
 pub mod event;
 pub mod persist;
 pub mod query;
+
+pub const LYNX_FORMAT_HEADER: &str = "X-Lynx-Format";
